\name{HypothTestforNonnormal}
\title{Assess the normality of of test statistcs when sample from nonnormal disttributions}
\description{A package to assess the normality of test statistcs when sample from common nonnormal disttributions}
\details{\tabular{11}{
Package:\tab HypothTestforNonnormal\cr
Type:\tab Package\cr
Version: \tab 0.1\cr
Date:\tab 2011-11-15\cr
License: \tab GPL (>=2)\cr
Lazyload: \tab yes\cr
}

how_norm functions simulate sample mean from given distribution and give QQ plot vs Normal}
\alias{HypothTestforNonnormal}
\docType{package}
\author{Qi Liu \email{qi.liu.1@vanderbilt.edu}}
\keyword{package}
\examples{how_norm.exp(n=5, rate=1, n_2=9, rate_2=1/2,sim=10^4)
how_norm.cauchy(n=1000,location=0, scale=1)}
